refactor(legacy): moved validation-rules into a rule folder, add type
