+++ src/utils/validation/rules/filter.ts
@@ -1,0 +1,7 @@
+import { required } from '../validate-rules';
+
+export const filterRules = {
+  initialMember: [required],
+  finalMember: [required],
+  filter: [required],
+};
