+++ src/utils/validation/rules/index.ts
@@ -1,0 +1,35 @@
+/*
+List of business rules to be applied to each questionnaire's components.
+*/
+import { calculatedVariableRules } from './calculatedVariable';
+import { collectedVariableRules } from './collectedVariable';
+import { controlRules } from './control';
+import { declarationRules } from './declaration';
+import { durationRulesPTnHnM, durationRulesPnYnM } from './duration';
+import { externalVariableRules } from './externalVariable';
+import { filterRules } from './filter';
+import { loopRules } from './loop';
+import { questionRules } from './question';
+import { questionnaireRules } from './questionnaire';
+import { redirectionRules } from './redirection';
+import { roundaboutRules } from './roundabout';
+import { sequenceRules } from './sequence';
+import { tableListMeasuresRules } from './tableListMeasures';
+
+export {
+  calculatedVariableRules,
+  collectedVariableRules,
+  controlRules,
+  declarationRules,
+  durationRulesPTnHnM,
+  durationRulesPnYnM,
+  externalVariableRules,
+  filterRules,
+  loopRules,
+  questionRules,
+  questionnaireRules,
+  redirectionRules,
+  roundaboutRules,
+  sequenceRules,
+  tableListMeasuresRules,
+};
