+++ src/utils/validation/rules/sequence.ts
@@ -1,0 +1,6 @@
+import { name, required } from '../validate-rules';
+
+export const sequenceRules = {
+  name: [required, name],
+  label: [required],
+};
